# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cellier API",
...
    "contact": {
      "name": "Equipo Cellier"
    },
    "license": {
      "name": "Uso interno"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/households": {
      "get": {
        "tags": [
          "Hogares"
        ],
        "summary": "Listar mis hogares",
        "description": "Los hogares a los que pertenece el usuario autenticado, con su rol en cada\nuno y el número de personas del grupo.\n\nEl orden es estable entre llamadas —por nombre, y por identificador a\nigualdad de nombre—, para que un selector de hogar en el cliente no se\nreordene solo. La lista no incluye el código de ingreso: ese dato sale\núnicamente por el detalle del hogar.\n",
        "operationId": "listMine",
        "responses": {
          "200": {
            "description": "Hogares del usuario. Lista vacía si aún no pertenece a ninguno.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseholdSummary"
                  }
                },
                "examples": {
                  "misHogares": {
                    "description": "misHogares",
                    "value": [
                      {
                        "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                        "name": "Casa Rivas",
                        "role": "ADMIN",
                        "memberCount": 3
                      },
                      {
                        "id": "b41d0f77-6c58-4e92-8a03-5fd91c2e7a64",
                        "name": "Depa Ñuñoa",
                        "role": "MEMBER",
                        "memberCount": 2
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Hogares"
        ],
        "summary": "Crear un hogar",
        "description": "Crea el hogar y deja a quien lo crea dentro como **administrador**, en la\nmisma operación. Un hogar sin administrador sería inadministrable, así que\nesa membresía no es un paso aparte que se pueda olvidar.\n\nEl código de ingreso se genera aquí y viaja en la respuesta.\n",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Hogar creado. La cabecera `Location` apunta a su detalle.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetail"
                },
                "examples": {
                  "hogarCreado": {
                    "description": "hogarCreado",
                    "value": {
                      "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                      "name": "Casa Rivas",
                      "joinCode": "K7M2QP9X",
                      "role": "ADMIN",
                      "memberCount": 1,
                      "createdAt": "2026-09-03T18:42:11Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "El nombre falta, está en blanco o pasa de 80 caracteres.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "nombreVacio": {
                    "description": "nombreVacio",
                    "value": {
                      "type": "https://cellier.app/problems/bad-request",
                      "title": "Datos inválidos",
                      "status": 400,
                      "detail": "Uno o más campos del cuerpo de la petición no son válidos.",
                      "instance": "/api/v1/households",
                      "errors": {
                        "name": "El nombre del hogar es obligatorio"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/households/{householdId}/join-requests/{joinRequestId}:reject": {
      "post": {
        "tags": [
          "Hogares"
        ],
        "summary": "Rechazar una solicitud",
        "description": "Deniega la entrada. No crea ninguna membresía y **no cierra la puerta para\nsiempre**: la persona puede volver a solicitarlo más adelante, porque el\nhueco de «una solicitud viva por hogar» solo lo ocupan las pendientes.\n\nRechazar no es lo mismo que cancelar: rechaza un administrador y queda\nconstancia de quién fue; cancelar solo puede hacerlo el propio solicitante.\n\nUna solicitud ya resuelta responde `409`.\n",
        "operationId": "reject",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "description": "Identificador del hogar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          },
          {
            "name": "joinRequestId",
            "in": "path",
            "description": "Identificador de la solicitud.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "5a1e93b7-2c04-4d88-91fe-77b6c0a2d431"
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud rechazada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequest"
                },
                "examples": {
                  "rechazada": {
                    "description": "rechazada",
                    "value": {
                      "id": "5a1e93b7-2c04-4d88-91fe-77b6c0a2d431",
                      "userId": "d90f4c21-77ab-4e35-9f62-1c8b03ad5e47",
                      "displayName": "Bruno Soto",
                      "email": "bruno.soto@gmail.com",
                      "avatarUrl": null,
                      "status": "REJECTED",
                      "requestedAt": "2026-09-04T09:15:02Z",
                      "resolvedAt": "2026-09-04T10:02:44Z",
                      "resolvedByUserId": "3f1c9a2e-5b7d-4a10-9c88-2f0e6b4d1a73"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/join-requests"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "El usuario es miembro del hogar, pero no administrador.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "noEsAdmin": {
                    "description": "noEsAdmin",
                    "value": {
                      "type": "https://cellier.app/problems/forbidden",
                      "title": "Operación no permitida",
                      "status": 403,
                      "detail": "Esta operación requiere rol de administrador en el hogar.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/join-requests"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "El hogar no es accesible para quien llama, o esa solicitud no pertenece a este hogar.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "solicitudDeOtroHogar": {
                    "description": "solicitudDeOtroHogar",
                    "value": {
                      "type": "https://cellier.app/problems/not-found",
                      "title": "Recurso no encontrado",
                      "status": 404,
                      "detail": "No existe esa solicitud en este hogar.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/join-requests/5a1e93b7-2c04-4d88-91fe-77b6c0a2d431:approve"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "La solicitud ya estaba resuelta.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "yaResuelta": {
                    "description": "yaResuelta",
                    "value": {
                      "type": "https://cellier.app/problems/conflict",
                      "title": "Conflicto con el estado actual",
                      "status": 409,
                      "detail": "Esa solicitud ya está resuelta.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/join-requests/5a1e93b7-2c04-4d88-91fe-77b6c0a2d431:approve"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/households/{householdId}/join-requests": {
      "get": {
        "tags": [
          "Hogares"
        ],
        "summary": "Listar las solicitudes del hogar",
        "description": "Las solicitudes dirigidas a este hogar, con los datos de quien las envía.\nSolo para administradores.\n\nSin `status` llegan todas, con las pendientes primero por ser las que piden\nuna decisión. Para la bandeja de trabajo habitual, `?status=PENDING`.\n\nCada solicitud incluye el correo del solicitante: aprobar da acceso a la\ndespensa de la casa, y el nombre para mostrar no basta para distinguir a\ndos homónimos.\n",
        "operationId": "list_4",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "description": "Identificador del hogar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtra por estado. Si se omite, llegan todas.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ]
            },
            "example": "PENDING"
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitudes del hogar.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JoinRequest"
                  }
                },
                "examples": {
                  "pendientes": {
                    "description": "pendientes",
                    "value": [
                      {
                        "id": "5a1e93b7-2c04-4d88-91fe-77b6c0a2d431",
                        "userId": "d90f4c21-77ab-4e35-9f62-1c8b03ad5e47",
                        "displayName": "Bruno Soto",
                        "email": "bruno.soto@gmail.com",
                        "avatarUrl": null,
                        "status": "PENDING",
                        "requestedAt": "2026-09-04T09:15:02Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/join-requests"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "El usuario es miembro del hogar, pero no administrador.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "noEsAdmin": {
                    "description": "noEsAdmin",
                    "value": {
                      "type": "https://cellier.app/problems/forbidden",
                      "title": "Operación no permitida",
                      "status": 403,
                      "detail": "Esta operación requiere rol de administrador en el hogar.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/join-requests"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "El hogar no existe, o existe pero el usuario autenticado no es miembro. La API no distingue ambos casos a propósito.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "hogarAjenoOInexistente": {
                    "description": "hogarAjenoOInexistente",
                    "value": {
                      "type": "https://cellier.app/problems/not-found",
                      "title": "Recurso no encontrado",
                      "status": 404,
                      "detail": "No existe ese hogar, o no eres miembro de él.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/join-requests"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "CreateHouseholdRequest": {
        "type": "object",
        "description": "Datos para crear un hogar.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del hogar.",
            "example": "Casa Rivas",
            "maxLength": 80,
            "minLength": 0
          }
        },
        "required": [
          "name"
        ]
      },
...
      "HouseholdSummary": {
        "type": "object",
        "description": "Hogar del usuario, con su rol y el tamaño del grupo.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador del hogar.",
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          },
          "name": {
            "type": "string",
            "description": "Nombre del hogar.",
            "example": "Casa Rivas"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario autenticado en este hogar.",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "example": "ADMIN"
          },
          "memberCount": {
            "type": "integer",
            "format": "int64",
            "description": "Número de personas que pertenecen al hogar.",
            "example": 3
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Token JWT emitido por Cellier.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}